Adds function for executing a program given a filename.
To make the execution of the interpreter as simple as possible, we
needed a function that could take a program name, looked for it and
executed it.

The function openAndExecuteProgram was added. This function takes a
filename and tries to open the file and execute the program from that file.